get started fix2

+++ app/main.py
@@ -36,11 +36,12 @@ def index():
 
 @app.route('/upload', methods=['GET', 'POST'])
 def upload():
-    for f in os.listdir(UPLOAD_FOLDER):
-        os.remove(os.path.join(UPLOAD_FOLDER, f))
+    if request.method == 'POST':
+        for f in os.listdir(UPLOAD_FOLDER):
+            os.remove(os.path.join(UPLOAD_FOLDER, f))
 
 
-    return render_template('index.html', start_bool = 0, upload_bool=1, classify_bool=0, try_another=0)
+        return render_template('index.html', start_bool = 0, upload_bool=1, classify_bool=0, try_another=0)
 
 
 @app.route('/classify', methods=['GET', 'POST'])
